+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [1.0.0] / 2024-09-05
 - Test Wall updater when the curve is changed.
+- Update `BuiltInParameterUpdater` to register `BuiltInParameter` array.
 
 [vNext]: ../../compare/1.0.0...HEAD
 [1.0.0]: ../../compare/1.0.0